move all fixtures as conftest.py plugin
